Move yet more node types out
